# contracts/openapi/Tims.Api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Tims.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/dei/dashboard-kpis": {
      "get": {
        "tags": [
          "Tims.Api"
        ],
        "operationId": "DeiGetDashboardKpis",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardKpis"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/dei/gender-representation": {
      "get": {
        "tags": [
          "Tims.Api"
        ],
        "operationId": "DeiGetGenderRepresentation",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenderRepresentationView"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/dei/age-distribution": {
      "get": {
        "tags": [
          "Tims.Api"
        ],
        "operationId": "DeiGetAgeDistribution",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgeDistributionView"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/dei/nationality-diversity": {
      "get": {
        "tags": [
          "Tims.Api"
        ],
        "operationId": "DeiGetNationalityDiversity",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NationalityDiversityView"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/dei/ethnicity-distribution": {
      "get": {
        "tags": [
          "Tims.Api"
        ],
        "operationId": "DeiGetEthnicityDistribution",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EthnicityDistributionView"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/dei/disability-distribution": {
      "get": {
        "tags": [
          "Tims.Api"
        ],
        "operationId": "DeiGetDisabilityDistribution",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisabilityDistributionView"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/dei/leadership-diversity": {
      "get": {
        "tags": [
          "Tims.Api"
        ],
        "operationId": "DeiGetLeadershipDiversity",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadershipDiversityResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/dei/hiring-funnel": {
      "get": {
        "tags": [
          "Tims.Api"
        ],
        "operationId": "DeiGetHiringFunnel",
        "parameters": [
          {
            "name": "dateFrom",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HiringFunnelView"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/dei/promotion-equity": {
      "get": {
        "tags": [
          "Tims.Api"
        ],
        "operationId": "DeiGetPromotionEquity",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromotionEquityView"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/dei/inclusion-index": {
      "get": {
        "tags": [
          "Tims.Api"
        ],
        "operationId": "DeiGetInclusionIndex",
        "parameters": [
          {
            "name": "surveyId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InclusionIndexResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/dei/pay-equity": {
      "get": {
        "tags": [
          "Tims.Api"
        ],
        "operationId": "DeiGetPayEquity",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayEquityView"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgeDistributionView": {
        "required": [
          "groups",
          "suppressed"
        ],
        "type": "object",
        "properties": {
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgeGroup"
            }
          },
          "suppressed": {
            "type": "boolean"
          }
        }
      },
      "AgeGroup": {
        "required": [
          "range",
          "count",
          "percentage",
          "suppressed"
        ],
        "type": "object",
        "properties": {
          "range": {
            "type": "string"
          },
          "count": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "percentage": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "suppressed": {
            "type": "boolean"
          }
        }
      },
...
      "DashboardKpis": {
        "required": [
          "totalEmployees",
          "demographicsCoverage",
          "genderParityIndex",
          "womenPct",
          "leadershipWomenPct",
          "totalNationalities"
        ],
        "type": "object",
        "properties": {
          "totalEmployees": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "demographicsCoverage": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "genderParityIndex": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "womenPct": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "leadershipWomenPct": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "totalNationalities": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "DisabilityDistributionView": {
        "required": [
          "groups",
          "suppressed"
        ],
        "type": "object",
        "properties": {
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DisabilityGroup"
            }
          },
          "suppressed": {
            "type": "boolean"
          }
        }
      },
...
      "EthnicityDistributionView": {
        "required": [
          "groups",
          "suppressed"
        ],
        "type": "object",
        "properties": {
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EthnicityGroup"
            }
          },
          "suppressed": {
            "type": "boolean"
          }
        }
      },
      "EthnicityGroup": {
        "required": [
          "ethnicity",
          "count",
          "percentage",
          "suppressed"
        ],
        "type": "object",
        "properties": {
          "ethnicity": {
            "type": "string"
          },
          "count": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "percentage": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "suppressed": {
            "type": "boolean"
          }
        }
      },
...
      "GenderGroup": {
        "required": [
          "gender",
          "count",
          "percentage",
          "suppressed"
        ],
        "type": "object",
        "properties": {
          "gender": {
            "type": "string"
          },
          "count": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "percentage": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "suppressed": {
            "type": "boolean"
          }
        }
      },
      "GenderRepresentationView": {
        "required": [
          "groups",
          "suppressed"
        ],
        "type": "object",
        "properties": {
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenderGroup"
            }
          },
          "suppressed": {
            "type": "boolean"
          }
        }
      },
...
      "HiringFunnelView": {
        "required": [
          "total"
        ],
        "type": "object",
        "properties": {
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "InclusionIndexResult": {
        "required": [
          "index",
          "totalResponses",
          "suppressed",
          "questionsEvaluated"
        ],
        "type": "object",
        "properties": {
          "index": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "totalResponses": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "suppressed": {
            "type": "boolean"
          },
          "questionsEvaluated": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "LeadershipDiversityResult": {
        "required": [
          "totalLeaders",
          "byGender",
          "suppressed"
        ],
        "type": "object",
        "properties": {
          "totalLeaders": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "byGender": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LeaderGroup"
            }
          },
          "suppressed": {
            "type": "boolean"
          }
        }
      },
...
      "NationalityDiversityView": {
        "required": [
          "totalNationalities",
          "distribution",
          "suppressed"
        ],
        "type": "object",
        "properties": {
          "totalNationalities": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "distribution": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NationalityGroup"
            }
          },
          "suppressed": {
            "type": "boolean"
          }
        }
      },
      "NationalityGroup": {
        "required": [
          "nationality",
          "count",
          "percentage",
          "suppressed"
        ],
        "type": "object",
        "properties": {
          "nationality": {
            "type": "string"
          },
          "count": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "percentage": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "suppressed": {
            "type": "boolean"
          }
        }
      },
...
      "PayEquityView": {
        "required": [
          "results",
          "gapPct",
          "suppressed",
          "currency"
        ],
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PayEquityGroup"
            }
          },
          "gapPct": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "suppressed": {
            "type": "boolean"
          },
          "currency": {
            "type": "string"
          }
        }
      },
...
      "PromotionEquityView": {
        "required": [
          "year",
          "totalPromotions",
          "suppressed"
        ],
        "type": "object",
        "properties": {
          "year": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalPromotions": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "suppressed": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}